requirements.txt
pyspark==3.5.0
numpy==1.26.0
pyarrow==15.0.0
findspark==2.0.1
python-dotenv==1.0.0
pymysql==1.1.0
sqlalchemy==2.0.25
scikit-learn==1.3.2
pandas==2.1.4
mysql-connector-python==8.2.0
streamlit==1.29.0
plotly==5.17.0
altair==5.2.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0